fix: Updated commands_test to work on *nix and Windows
