+++ internal/inputs/commands_test.go
@@ -1,5 +1,3 @@
-//+build linux darwin
-
 /*
 * Copyright 2019 New Relic Corporation. All rights reserved.
 * SPDX-License-Identifier: Apache-2.0
@@ -154,19 +152,6 @@ func TestCanRunMultipleCommands(t *testing.T) {
 	}
 }
 
-func getCanRunMultipleCommands() []load.Command {
-	return []load.Command{
-		{
-			Run:     "cat ../../test/payloads/redisInfo.out",
-			SplitBy: ":",
-		},
-		{
-			Run:     `echo "zHost:$(echo HELLO)"`,
-			SplitBy: ":",
-		},
-	}
-}
-
 func TestDf(t *testing.T) {
 	load.Refresh()
 	config := load.Config{
@@ -203,28 +188,6 @@ func TestDf(t *testing.T) {
 	}
 }
 
-func getDfApis() []load.API {
-	return []load.API{
-		{
-			Name:  "df",
-			Shell: "/bin/sh",
-			Commands: []load.Command{
-				{
-					Run:      "cat ../../test/payloads/df.out",
-					Split:    "horizontal",
-					RowStart: 1,
-					SetHeader: []string{
-						"fs", "512Blocks", "used", "available", "capacity", "iused", "ifree", "iusedPerc", "mountedOn",
-					},
-					RegexMatch: false,
-					SplitBy:    `\s{1,}`,
-					Shell:      "/bin/sh",
-				},
-			},
-		},
-	}
-}
-
 func TestDf2(t *testing.T) {
 	load.Refresh()
 	config := load.Config{
@@ -258,23 +221,5 @@ func TestDf2(t *testing.T) {
 	for key, expectedValue := range expected {
 		actualValue := actual[key]
 		assert.Equalf(t, expectedValue, actualValue, "%s doesnt match - want: %v  got: %v", key, expectedValue, actualValue)
-	}
-}
-
-func getDf2Apis() []load.API {
-	return []load.API{
-		{
-			Name: "df",
-			Commands: []load.Command{
-				{
-					Run:              "cat ../../test/payloads/df.out",
-					Split:            "horizontal",
-					RegexMatch:       true,
-					SplitBy:          `(\S+)\s+(\d+)\s+(\d+)\s+(\d+)\s+(\d+)%\s+(\d+)\s+(\d+)\s+(\d+)%\s+(.*)`,
-					HeaderRegexMatch: false,
-					HeaderSplitBy:    `\s{1,}`,
-				},
-			},
-		},
 	}
 }
